backend/requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Redis
redis==5.0.1
aioredis==2.0.1

# Trading
ccxt==4.1.22
pandas==2.1.3
numpy==1.26.2

# WebSocket
websockets==12.0
python-socketio==5.10.0

# Monitoring
prometheus-client==0.19.0

# API Documentation
pydantic==2.5.0
pydantic-settings==2.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Background Tasks
celery==5.3.4
flower==2.0.1

# Utils
python-json-logger==2.0.7
tenacity==8.2.3